+++ src/utils.ts
@@ -2,6 +2,9 @@ import { ActorPF2e, TokenPF2e, MeasuredTemplateDocumentPF2e, ItemPF2e, Condition
 import { getSetting, SETTINGS } from "./settings.ts";
 import { MeasuredTemplateType } from "foundry-pf2e/foundry/common/constants.mjs";
 import { Point } from "foundry-pf2e/foundry/common/_types.mjs";
+import { TokenMovementMethod } from "foundry-pf2e/foundry/client/_types.mjs";
+
+export type Tradition = "occult" | "arcane" | "divine" | "primal";
 
 export const MODULE_ID = "samioli-module";
 
@@ -316,4 +319,54 @@ export function getTokensOnCurrentSceneForActor(actor: ActorPF2e): TokenDocument
     const tokens = currentScene.tokens.filter(t => t.actorId === actor.id);
 
     return tokens;
+}
+
+export async function moveTokenToPoint(token: TokenPF2e, point: Point, ignoreWalls?: boolean, ignoreCost?: boolean) {
+
+    const waypoints = [{
+        x: point.x,
+        y: point.y
+    }];
+
+    const moveOptions = {
+        method: "api" as TokenMovementMethod, 
+        autoRotate: false,
+        showRuler: false,
+        constrainOptions: {
+            ignoreWalls: ignoreWalls,
+            ignoreCost: ignoreCost
+        }
+    };
+
+    await token.document.move(waypoints, moveOptions);
+}
+
+/**
+ * Returns all tokens at a given location, except for loot and party tokens.
+ */
+export function getTokensAtLocation(location: Point, includeHidden?: boolean): TokenPF2e[] {
+
+    const locationGridOffset = canvas.grid.getOffset(location);
+    const allTokensOnScene = canvas.tokens.placeables;
+    
+    const validTokens = allTokensOnScene.filter(token => {
+        
+        const actorType = token.actor?.type;
+        if (actorType === "loot" || actorType === "party") {
+            return false;
+        }
+        return token.footprint.some(footprint => {
+            return footprint.i === locationGridOffset.i && footprint.j === locationGridOffset.j;
+        });
+    }) as TokenPF2e[];
+
+    if (includeHidden) {
+        return validTokens;
+    }
+
+    const visibleTokens = validTokens.filter(token => { 
+        return token.document.hidden === false;
+    });
+
+    return visibleTokens;
 }